+++ src/main/java/de/adorsys/keycloak/config/service/RealmImportService.java
@@ -56,6 +56,8 @@ public class RealmImportService {
             "requiredActions",
             "defaultDefaultClientScopes",
             "defaultOptionalClientScopes",
+            "clientProfiles",
+            "clientPolicies",
     };
 
     private static final Logger logger = LoggerFactory.getLogger(RealmImportService.class);
@@ -64,6 +66,9 @@ public class RealmImportService {
 
     private final UserImportService userImportService;
     private final UserProfileImportService userProfileImportService;
+
+    private final ClientPoliciesImportService clientPoliciesImportService;
+
     private final RoleImportService roleImportService;
     private final ClientImportService clientImportService;
     private final ClientScopeImportService clientScopeImportService;
@@ -90,6 +95,7 @@ public class RealmImportService {
             RealmRepository realmRepository,
             UserImportService userImportService,
             UserProfileImportService userProfileImportService,
+            ClientPoliciesImportService clientPoliciesImportService,
             RoleImportService roleImportService,
             ClientImportService clientImportService,
             GroupImportService groupImportService,
@@ -110,6 +116,7 @@ public class RealmImportService {
         this.realmRepository = realmRepository;
         this.userImportService = userImportService;
         this.userProfileImportService = userProfileImportService;
+        this.clientPoliciesImportService = clientPoliciesImportService;
         this.roleImportService = roleImportService;
         this.clientImportService = clientImportService;
         this.groupImportService = groupImportService;
@@ -194,6 +201,7 @@ public class RealmImportService {
         defaultGroupsImportService.doImport(realmImport);
         componentImportService.doImport(realmImport);
         userProfileImportService.doImport(realmImport);
+        clientPoliciesImportService.doImport(realmImport);
         userImportService.doImport(realmImport);
         requiredActionsImportService.doImport(realmImport);
         authenticationFlowsImportService.doImport(realmImport);