+++ src/test/java/de/adorsys/keycloak/config/AbstractImportIT.java
@@ -74,7 +74,7 @@ abstract public class AbstractImportIT extends AbstractImportTest {
             KEYCLOAK_CONTAINER.setCommand("start-dev");
             command.add("start-dev");
             command.add("--features");
-            command.add("admin-fine-grained-authz,declarative-user-profile");
+            command.add("admin-fine-grained-authz,declarative-user-profile,client-policies,client-secret-rotation");
         }
 
         if (System.getProperties().getOrDefault("skipContainerStart", "false").equals("false")) {